requirements.txt
Django
psycopg2-binary
python-dotenv
djangorestframework